pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Andrey M. <tolkoigri@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
django-bootstrap5 = "^24.3"
flake8 = "^7.1.1"
gunicorn = "^23.0.0"
django-filter = "^24.3"
rollbar = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
